Added 1.9 -> 1.9 and 1.12 -> 1.12 id/meta blockstate mappings, remove 1.12 -> 1.18.10 blockstate mappings
These new mappings are much more useful, as they allow much more fine-grained upgrading.
Previously, it was only possible to upgrade a <= 1.12 world's blocks directly to 1.18.10, as that's the earliest version of blockstate mapping I could find.

These new mappings allow translating id/meta to blockstate NBT within the version itself. This means that you can directly translate, for example, 1.12 id/meta to 1.12 NBT without any other unrelated upgrades.

The 1.12.0 mapping can be used for the same purpose as the old 1.18.10 mapping.
The 1.9.0 mapping won't be needed for most cases, but may be needed if the user wants to upgrade to a version in the range 1.9-1.11, as the 1.12 mapping won't be suitable for upgrading blockitems in inventories for such data.

By virtue of this, some old bugs have also been fixed:
- blast_furnace and smoker placed in 1.10 now upgrade their facing values correctly
- cocoa age being invalid no longer causes loss of the facing property when upgraded
- plus whatever other bugs were fixed by the upgrade schemas that weren't fixed by Mojang